Move ZKPP files into libchelonia
Closes #2864
