--- shared/zkppConstants.js
@@ -1,6 +1,0 @@
-export const AUTHSALT = 'AUTHSALT'
-export const CONTRACTSALT = 'CONTRACTSALT'
-export const CS = 'CS'
-export const SU = 'SU'
-
-export const SALT_LENGTH_IN_OCTETS = 24
